refactor: decouple logic into services and implement Pokemon build management with Supabase
