+++ backend/routers/damage.py
@@ -1,36 +1,24 @@
 from fastapi import APIRouter
-from pydantic import BaseModel
-import math
+from schemas.damage import DamageRequest, DamageResponse
+from services.damage import calculate_pokemon_damage
 
-# ルーターの作成（main.py の app の代わりになります）
 router = APIRouter()
 
-# リクエストモデル
-class DamageRequest(BaseModel):
-    level: int
-    power: int
-    attack: int
-    defense: int
-    is_stab: bool
-    effectiveness: float
-
-# エンドポイントの定義（@app.post ではなく @router.post になります）
-@router.post("/")
+@router.post("/", response_model=DamageResponse)
 def calculate_damage(req: DamageRequest):
-    # --- ここは以前書いた計算ロジックをそのままペースト ---
-    step1 = math.trunc(2 * req.level / 5) + 2
-    step2 = math.trunc(step1 * req.power * req.attack / req.defense)
-    base_damage = math.trunc(step2 / 50) + 2
-    
-    if req.is_stab:
-        base_damage = math.trunc(base_damage * 1.5)
-        
-    base_damage = math.trunc(base_damage * req.effectiveness)
-    
-    min_damage = math.trunc(base_damage * 85 / 100)
-    max_damage = math.trunc(base_damage * 100 / 100)
+    """
+    ダメージ計算を行うエンドポイント
+    """
+    min_dmg, max_dmg = calculate_pokemon_damage(
+        level=req.level,
+        power=req.power,
+        attack=req.attack,
+        defense=req.defense,
+        is_stab=req.is_stab,
+        effectiveness=req.effectiveness
+    )
     
     return {
-        "min_damage": min_damage,
-        "max_damage": max_damage
+        "min_damage": min_dmg,
+        "max_damage": max_dmg
     }
